create forward declaration header library
